requirements.txt
telethon
google-generativeai
python-dotenv